+++ resources/views/partials/sidebar.blade.php
@@ -258,6 +258,7 @@
         $biebMenu['Bibliotheek'][] = ['Publicatie toevoegen', 'bibliotheek.publicaties.create', 'plus', 'bibliotheek.publicaties.create'];
         $biebMenu['Bibliotheek'][] = ['Uitlenen', 'bibliotheek.uitlenen', 'taak', 'bibliotheek.uitlenen'];
         $biebMenu['Bibliotheek'][] = ['Importeren', 'bibliotheek.import', 'db', 'bibliotheek.import'];
+        $biebMenu['Bibliotheek'][] = ['Verrijking (ISBN)', 'bibliotheek.verrijking', 'search', 'bibliotheek.verrijking'];
     }
     if ($gebruiker->magBibliotheekSjablonenBeheren()) {
         $biebMenu['Bibliotheek'][] = ['E-mailsjablonen', 'bibliotheek.sjablonen', 'log', 'bibliotheek.sjablonen'];